requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pillow==9.5.0
torch==2.6.0
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cublas-cu12==12.4.5.8
faiss-gpu-cu12==1.10.0
git+https://github.com/openai/CLIP.git
python-multipart
